+++ previous-versions/files/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.1.1] previous-versions/files
 ### Releases
 - [PyPI v0.1.1] https://pypi.org/project/codat-files/0.1.1 - previous-versions/files
+
+## 2023-09-29 00:01:00
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.0] previous-versions/files
+### Releases
+- [PyPI v0.2.0] https://pypi.org/project/codat-files/0.2.0 - previous-versions/files